+++ analyse/src/mitos_read_2d.c
@@ -28,10 +28,10 @@
 #include "../../include/mitos.h"
 
 
-static inline int row_major_2d(int i, int j, int N) {
+static inline long long int row_major_2d(int i, int j, int N) {
     i = wrap(i,N);
     j = wrap(j,N);
-    return i*N + j;
+    return (long long int) i*N + j;
 }
